rearranged deps for utils.py and location of safe_profile_select to make more sense. As utils.py no required models.py as a dependency, it has to be loaded only after the app is initialized
