+++ stylometryproject/stylometryapp/models.py
@@ -1,6 +1,5 @@
 from django.db import models
 from django.contrib.auth.models import User
-from django.http import HttpRequest
 
 class Profile(models.Model):
     """ Profile Table """
@@ -22,26 +21,3 @@ class Document(models.Model):
     
     def __str__(self):
         return self.title
-
-# Non-class function to check profile stuff
-def safe_profile_select(request: HttpRequest, profile:Profile = None) -> Profile|None:
-    """Allows querying and selection of current profile based on a request
-    Returns the currently selected Profile or None, but only if it
-    exists, is a valid profile and belongs to the user asking for it"""
-
-    # Set the profile to operate on
-    profile = request.session.get('profile_cur') if not profile else profile
-    if not isinstance(profile, Profile): return None  # No valid reference to a current profile anywhere
-
-    # Catch when a profile doesn't exist or doesn't belong to the user
-    if not Profile.objects.filter(user=request.user, id=profile.id).exists():
-        profile = None
-    
-    # No profile? Remove the session variable
-    if profile:
-        request.session['profile_cur'] = profile
-    else:
-        if 'profile_cur' in request.session:
-            request.session.pop('profile_cur')
-    
-    return profile  # Will be None if profile doesn't pass
